src/requirements.txt
snowflake-snowpark-python==1.16.0
PyPDF2==2.10.5
langchain==0.0.298